MYC-13 update server structure for tests
