+++ src/types/Transaction.ts
@@ -137,19 +137,11 @@ export class Transaction {
         const senderAddress = body.sender ?? address(body.senderPublicKey, CHAIN_ID);
         const sponsor = await this.getSponsor(senderAddress, this?.block?.getHeight());
 
-        if(this.block) {
-            return Promise.resolve([
-                Operation.create(0, body.recipient, body.amount, OperationTypes.Transfer),
-                Operation.create(1, senderAddress, -body.amount, OperationTypes.Transfer),
-                Operation.create(2, sponsor ?? senderAddress, -body.fee, OperationTypes.Transfer)
-            ]);
-        } else {
             return Promise.resolve([
                 Operation.createNew(0, body.recipient, body.amount, OperationTypes.Transfer),
                 Operation.createNew(1, senderAddress, -body.amount, OperationTypes.Transfer),
                 Operation.createNew(2, sponsor ?? senderAddress, -body.fee, OperationTypes.Transfer)
             ]);
-        }
 
     }
